requirements.txt
Click==7.0
urllib3==1.24.3
requests==2.23.0
jsonpath-rw==1.4.0
PyYAML==5.3.1
ruamel.yaml==0.16.12
Flask==1.0.2
Flask-RESTful==0.3.7
flasgger==0.9.2
gevent==1.4.0
greenlet==0.4.16
importlib-resources==1.0.2
APScheduler==3.6.3
